fix baccaml_cli and stub
